requirements.txt
autopep8==1.3.4
Django==2.0.2
django-adminlte2==0.3.0
djangorestframework==3.8.2
mysql-connector-python==8.0.6
Pillow==5.0.0
pkg-resources==0.0.0
pycodestyle==2.3.1
pytz==2018.3
